requirements.txt
mysql-connector-python
Pillow
bcrypt
flask
python-dotenv
